requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

requests
beautifulsoup4
lxml
trafilatura
playwright

sentence-transformers
faiss-cpu
numpy
pandas
openpyxl

streamlit
httpx
